requirements.txt
python-magic-bin
python-nmap
scapy
rich